app.py
import streamlit as st
import numpy as np
import cv2
from PIL import Image

# Assuming you have custom modules for enhancement and detection
# from models.enhancement import enhance_image
# from models.detection import detect_objects

def enhance_image(image):
    # Placeholder: identity function
    return image

def detect_objects(image, model_type="fish"):
    # Placeholder: draw dummy boxes
    height, width, _ = image.shape
    cv2.rectangle(image, (50, 50), (width - 50, height - 50), (0, 255, 0), 3)
    cv2.putText(image, f"{model_type.capitalize()} Detected", (60, 70),
                cv2.FONT_HERSHEY_SIMPLEX, 1, (0, 255, 0), 2)
    return image

# Streamlit UI
st.title("Underwater Image Enhancement & Detection")

uploaded_file = st.file_uploader("Upload an underwater image", type=["jpg", "jpeg", "png"])

if uploaded_file:
    image = Image.open(uploaded_file).convert("RGB")
    image_np = np.array(image)
    st.image(image_np, caption="Original Image", use_column_width=True)

    if st.button("Enhance Image"):
        enhanced_image = enhance_image(image_np)
        st.image(enhanced_image, caption="Enhanced Image", use_column_width=True)
        
        detection_type = st.selectbox("Select Detection Type", ["fish", "coral"])
        detected_image = detect_objects(enhanced_image.copy(), model_type=detection_type)
        st.image(detected_image, caption=f"{detection_type.capitalize()} Detection", use_column_width=True)
