requirements.txt
streamlit==1.35.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python
Pillow
numpy
matplotlib
scikit-image
